fix test fixture for mec
