+++ tests/integration/configs/mec_small.yaml
@@ -2,9 +2,17 @@
 description: |
   Integration-test fixture: minimal --mec run exercising the MEC feedback-file
   path (prepare_mec_input -> link_mec_input -> generate_mec_namelist -> run_mec).
-  Inference GRIB is pre-staged in output_root so no GPU/MLflow is needed.
-  Driven by tests/integration/test_mec.py.
+  Inference is replayed from a frozen fixture (no GPU / MLflow / venv build); the
+  verSYNOP obs are frozen alongside it under the same fixture dir.
+  Driven by tests/integration/test_mec_small.py.
 config_label: mec-test
+
+# Always replay inference from this frozen fixture (no GPU / MLflow / sandbox
+# build). To (re)capture the fixture you must run real inference, so temporarily
+# comment this line out, run `evalml experiment <this config> --mec`, then
+# `evalml capture-fixture <this config> <path>`, and restore it (see
+# tests/integration/README.md).
+fixture_root: /store_new/mch/msopr/ml/evalml/test_fixtures/mec-small/v_20250801_9f46
 
 dates:
   start: 2025-08-01T12:00
@@ -37,14 +45,15 @@ experiment:
       - season
 
 locations:
-  output_root: /store_new/mch/msopr/ml/evalml/test_fixtures/mec-small/
+  output_root: output/
 
 # Required for --mec: prepare_mec_input reads observations from these three roots.
-# Only ver_synop_root is used by this test.
+# Only ver_synop_root is used by this test; it points at the frozen verSYNOP obs
+# co-located with the inference fixture (see fixture_root above).
 mec:
   ekf_root: /store_new/mch/msopr/osm/KENDA-CH1/EKF
   mon_synop_root: /scratch/mch/paa/mec/MEC_ML_input/monFiles2025
-  ver_synop_root: /store_new/mch/msopr/ml/evalml/test_fixtures/mec-small/data/runs/forecaster-9f46-a697/6ec6/obs
+  ver_synop_root: /store_new/mch/msopr/ml/evalml/test_fixtures/mec-small/v_20250801_9f46/obs
 
 profile:
   executor: slurm
