+++ workflow/rules/verif_obs.smk
@@ -80,12 +80,11 @@ if config["mec"] is not None:
         This assembles the model input directory that MEC expects for a single valid time.
         """
         input:
-            # depend on ALL source grib dirs: for each lead l, source_init = init_time - l hours
             obs_file=rules.prepare_mec_input.output.obs_file,
-            src_dirs=lambda wc: expand(
-                str(OUT_ROOT / "data/runs/{run_id}/{src_init}/grib"),
+            inference_ok=lambda wc: expand(
+                rules.inference_execute.output.okfile,
                 run_id=wc.run_id,
-                src_init=[
+                init_time=[
                     (
                         datetime.strptime(wc.init_time, "%Y%m%d%H%M")
                         - timedelta(hours=l)
@@ -249,7 +248,9 @@ if config["mec"] is not None:
                     OUT_ROOT / f"data/runs/{wc.run_id}/fdbk_files"
                 ),
                 # Keeping this as a param. We will create it in run_ffv2 rule.
-                output_directory=lambda wc: str(OUT_ROOT / f"data/runs/{wc.run_id}/scores"),
+                output_directory=lambda wc: str(
+                    OUT_ROOT / f"data/runs/{wc.run_id}/scores"
+                ),
                 # TODO: consider including run_ids here?
                 experiment_ids=config["ffv2"]["experiment_ids"],
                 veri_ens_member=config["ffv2"]["veri_ens_member"],
@@ -287,7 +288,7 @@ if config["mec"] is not None:
             localrule: True
             shell:
                 "sarus pull container-registry.meteoswiss.ch/ffv2ctr/ffv2-container:0.1.0-main"
-    
+
         rule run_ffv2:
             input:
                 namelist=rules.generate_ffv2_namelist.output.namelist,
@@ -319,10 +320,10 @@ if config["mec"] is not None:
                 (
                     set -euo pipefail
                     echo "...time at start of run_ffv2: $(date)"
-    
+
                     # Create the output directory to hold scores, if it does not exist
                     mkdir -p {output.scores}
-    
+
                     namelist=$(realpath {input.namelist})
                     domain_table={params.domain_table}
                     blacklists={params.blacklists}
@@ -336,11 +337,11 @@ if config["mec"] is not None:
                         --mount=type=bind,source=$feedback_dir_abs,destination=/src/ffv2/input \
                         --mount=type=bind,source=$output_dir_abs,destination=/src/ffv2/output \
                         container-registry.meteoswiss.ch/ffv2ctr/ffv2-container:0.1.0-main
-    
+
                     echo "...time at end of run_ffv2: $(date)"
                 ) >{log} 2>&1
                 """
-    
+
         rule reorganize_ffv2_files:
             input:
                 scores=rules.run_ffv2.output.scores,
@@ -354,10 +355,10 @@ if config["mec"] is not None:
                 (
                     set -euo pipefail
                     echo "...time at start of reorganize_ffv2_files: $(date)"
-    
+
                     input_dir_abs=$(realpath {input.scores})
                     output_dir_abs=$(realpath {output.shiny_dir})
-    
+
                     # move score files into app-specific subdirectories, for the Shiny app
                     # display.
                     mkdir -p $output_dir_abs/fdbk_cont/data
@@ -366,25 +367,25 @@ if config["mec"] is not None:
                     mkdir -p $output_dir_abs/fdbk_synop_categ/data
                     mkdir -p $output_dir_abs/fdbk_synop_categ_bystat/data
                     mkdir -p $output_dir_abs/fdbk_synop_categ_ts/data
-    
+
                     # DET surface continuous scores
                     cp $input_dir_abs/CONT_exp* $output_dir_abs/fdbk_cont/data/
                     # DET surface continuous scores as time series
                     cp $input_dir_abs/CONT_TS_exp* $output_dir_abs/fdbk_cont_ts/data/
                     # DET surface continuous by stations
                     cp $input_dir_abs/CONT_bs_exp* $output_dir_abs/fdbk_cont_bystat/data/
-    
+
                     # Categorical verification against SYNOP
                     cp $input_dir_abs/CATEG_exp* $output_dir_abs/fdbk_synop_categ/data
                     cp $input_dir_abs/PEC_exp* $output_dir_abs/fdbk_synop_categ/data
-    
+
                     # Categorical verification against SYNOP by station
                     # This is not presently generated, so skip.
                     #cp $input_dir_abs/CATEG_TS_exp* $output_dir_abs/fdbk_synop_categ_ts/data
-    
+
                     # Categorical verification against SYNOP as time series
                     cp $input_dir_abs/CATEG_bs_exp* $output_dir_abs/fdbk_synop_categ_bystat/data
-    
+
                     echo "...time at end of reorganize_ffv2_files: $(date)"
                 ) >{log} 2>&1
                 """
